Introduce icon mapping to avoid duplicates
This commit adds a semantic icon library and a mapping configuration to avoid duplicate icons in icon packs.

The semnatic icon library contains an entry for every unique icon and a description of it's purpose. The mapping configuration maps every unique icon to all currently existing eclipse icons that should be replaced by the unique icon.